requirements.txt
pinecone-client[grpc]==2.2.1
sentence-transformers==2.2.2
datasets
apache-beam
protobuf==3.20.0
torch
streamlit-chat-media
streamlit-chat
transformers
